backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
supabase
python-dotenv
